+++ src/win.h
@@ -81,6 +81,7 @@ extern void win_set_font_size(int, bool sync_size_with_font);
 extern uint win_get_font_size(void);
 
 extern bool dw_check_glyphs(xchar * xcs, uint num, cattrflags attr);
+extern bool dw_has_glyph(xchar xc, cattrflags attr);
 extern void win_check_glyphs(wchar *wcs, uint num, cattrflags attr);
 extern wchar get_errch(wchar *wcs, cattrflags attr);
 extern int win_char_width(xchar, cattrflags attr);